Extracted title mapping to hook
